#19 opening/kindtypecomponent 리펙토링중..
