+++ src/components/opening/KindTypeComponent/svg/DebateComponent.tsx
@@ -1,0 +1,46 @@
+import { useRecoilState } from "recoil";
+import { studyFormState } from "src/components/states/studyForm";
+
+export default function AnnouncementComponent(): JSX.Element {
+  const [studyForm, setStudyForm] = useRecoilState(studyFormState);
+
+  return (
+    <>
+      {studyForm.type === "announcement" ? (
+        <>
+          <g filter="url(#filter2_d)">
+            <circle
+              cursor="pointer"
+              onClick={() => setStudyForm({ ...studyForm, type: "" })}
+              cx="68"
+              cy="160"
+              r="55"
+              fill="#9A6EC6"
+            />
+            <path
+              cursor="pointer"
+              onClick={() => setStudyForm({ ...studyForm, type: "" })}
+              d="M61.073 159.56H57.517V157.684H61.073V159.56ZM62.193 154.996H61.073V156.802H57.517V154.996H56.383V160.484H62.193V154.996ZM58.791 164.614H65.609V161.562H57.629V162.472H64.475V163.76H57.671V166.896H66.071V165.986H58.791V164.614ZM65.609 157.18V154.492H64.461V160.974H65.609V158.132H67.471V157.18H65.609ZM73.042 161.59H75.772V164.572H73.042V161.59ZM72.846 156.564H75.968V160.666H72.846V156.564ZM76.92 164.572V161.59H79.048V160.666H77.116V156.564H79.062V155.64H69.752V156.564H71.698V160.666H69.78V161.59H71.894V164.572H68.73V165.51H80.126V164.572H76.92Z"
+              fill="#222222"
+            />
+          </g>
+        </>
+      ) : (
+        <>
+          <path
+            cursor="pointer"
+            onClick={() => setStudyForm({ ...studyForm, type: "write" })}
+            d="M149.263 89.6411C149.263 74.2451 165.929 64.6225 179.263 72.3205L284.263 132.942C297.596 140.64 297.596 159.885 284.263 167.583L179.263 228.205C165.93 235.903 149.263 226.281 149.263 210.885L149.263 89.6411Z"
+            fill="#D2F3E6"
+          />
+          <path
+            cursor="pointer"
+            onClick={() => setStudyForm({ ...studyForm, type: "write" })}
+            d="M188.444 154.502H195.892V151.338H187.268V152.248H194.744V153.634H187.31V156.882H196.2V155.972H188.444V154.502ZM195.57 149.154C195.85 147.74 195.85 146.662 195.85 145.808V145.01H187.38V145.934H194.716C194.716 146.76 194.702 147.768 194.422 149.154H185.924V150.078H197.32V149.154H195.57ZM210.045 150.904C208.393 150.204 207.525 148.622 207.525 147.096V145.5H206.363V147.096C206.363 148.328 205.747 149.84 204.487 150.708C203.213 149.812 202.569 148.258 202.569 147.096V145.5H201.435V147.096C201.435 148.482 200.525 150.176 198.915 150.904L199.531 151.786C200.721 151.212 201.561 150.148 201.995 148.958C202.387 150.176 203.213 151.338 204.473 151.926C205.733 151.352 206.545 150.232 206.951 149.028C207.371 150.218 208.211 151.24 209.415 151.786L210.045 150.904ZM198.803 154.418V155.37H210.185V154.418H198.803ZM212.423 145.836V146.76H217.141C216.903 149.742 215.223 152.15 211.849 153.76L212.465 154.698C216.637 152.654 218.303 149.462 218.303 145.836H212.423ZM220.837 144.478V157.036H221.985V144.478H220.837Z"
+            fill="#222222"
+          />
+        </>
+      )}
+    </>
+  );
+}
